requirements.txt
flask==2.0.1
python-dotenv==0.19.0
openai==1.3.0
moviepy==1.0.3
pillow==9.0.0
pymongo==4.1.1
python-whatsapp-sdk==0.1.0
requests==2.26.0
gunicorn==20.1.0
python-multipart==0.0.5 